Make improvements
- Improved async signal safety of read() particularly for longjmp()
- Started adding cancel cleanup handlers for locks / etc on Windows
- Make /dev/tty work better particularly for uses like `foo | less`
- Eagerly read console input into a linked list, so poll can signal
- Fix some libc definitional bugs, which configure scripts detected

+++ libc/sock/struct/pollfd.internal.h
@@ -12,7 +12,7 @@ int32_t __sys_poll(struct pollfd *, uint64_t, signed);
 int sys_ppoll(struct pollfd *, size_t, const struct timespec *,
               const sigset_t *, size_t);
 int sys_poll_metal(struct pollfd *, size_t, unsigned);
-int sys_poll_nt(struct pollfd *, uint64_t, uint64_t *, const sigset_t *);
+int sys_poll_nt(struct pollfd *, uint64_t, uint32_t *, const sigset_t *);
 
 const char *DescribePollFds(char[300], ssize_t, struct pollfd *, size_t);
 #define DescribePollFds(x, y, z) DescribePollFds(alloca(300), x, y, z)
